ai_blog_agent/requirements.txt
langchain
langchain-openai
langchain-community
openai
python-dotenv
requests
PyYAML
python-telegram-bot 